pyproject.toml
[project]
name = "rpa-harness"
version = "0.1.0"
description = "AI-powered RPA automation harness with Playwright, Windows UIAutomation, and agentic AI loop"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RPA Harness" }]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "aiofiles>=23.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "pyautogui>=0.9.54",
    "python-docx>=1.0.0",
    "pypdf>=4.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
windows = [
    "pywinauto>=0.6.8",
    "pywin32>=306",
    "comtypes>=1.4.0",
]
outlook = [
    "pywin32>=306",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-playwright>=0.5.0",
]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[project.urls]
Repository = ""

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
